Prevent teleporting while riding a vehicle
The TeleportEvent has been updated to prevent players from teleporting while riding or seated in a vehicle. Instead of dismounting the player and adjusting the destination's coordinates, the request is now simply cancelled and a message is sent to the player, explaining why the teleportation was not completed. This change improves the clarity of the situation for the player and fixed the glitch that lets players teleport into places they shouldn't be.